+++ inc/functions/security-assessment.php
@@ -5,6 +5,27 @@ use PhpOffice\PhpSpreadsheet\IOFactory;
 use PhpOffice\PhpSpreadsheet\Spreadsheet;
 
 function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
+    // ** Stuff that may change when the template does ** //
+    // New slides to be appended after this slide number
+    $ThreatActorSlideStart = 32;
+    // The Threat Actor PNG ID - Found in slideX.xml.rels
+    $ThreatActorPNGID = 'rId20';
+    // The Threat Actor SVG ID - Found in slideX.xml.rels
+    $ThreatActorSVGID = 'rId21';
+    // Recognised Threat Actor List for image generation
+    $KnownActors = array(
+        'Venal Viper' => 'vextrio-viper', // Needs updating to Venal Viper image
+        'Vigorish Viper' => 'vextrio-viper', // Needs updating to Vigorish Viper image
+        'VexTrio' => 'vextrio-viper',
+        'Loopy Lizard' => 'loopy-lizard',
+        'Decoy Dog' => 'decoy-dog',
+        'Savvy Seahorse' => 'savvy-seahorse',
+        'Prolific Puma' => 'prolific-puma',
+        'Revolver Rabbit' => 'revolver-rabbit-noglow', // Needs updating to one with green glow
+        'Muddling Meerkat' => 'muddling-meerkat'
+    );
+    // ** Stuff that may change when the template does ** //
+
     // Check API Key is valid & get User Info
     $UserInfo = GetCSPCurrentUser();
     if (!isset($UserInfo->Error)) {
@@ -13,24 +34,18 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
         $StartDimension = str_replace('Z','',$StartDateTime);
         $EndDimension = str_replace('Z','',$EndDateTime);
 
-        // Set Directory
+        // Set Directories
         $FilesDir = __DIR__.'/../../files';
-
-        // Extract Powerpoint Template Strings
-        // ** Using external library to save re-writing the string replacement functions manually. Will probably pull this in as native code at some point.
-        $extractor = new BasicExtractor();
-        $mapping = $extractor->extractStringsAndCreateMappingFile(
-            $FilesDir.'/template-sept-24.pptx',
-            $FilesDir.'/reports/report-'.$UUID.'-extracted.pptx'
-        );
-        $Progress = writeProgress($UUID,$Progress);
+        $AssetsDir = __DIR__.'/../../assets';
 
         // Extract Powerpoint Template Zip
-        extractZip($FilesDir.'/reports/report-'.$UUID.'-extracted.pptx',$FilesDir.'/reports/report-'.$UUID);
-        $Progress = writeProgress($UUID,$Progress);
+        $Progress = writeProgress($UUID,$Progress,"Extracting template");
+        extractZip($FilesDir.'/template-threat-actors.pptx',$FilesDir.'/reports/report-'.$UUID);
+
         //
         // Do Chart, Spreadsheet & Image Stuff Here ....
         // Top threat feeds
+        $Progress = writeProgress($UUID,$Progress,"Getting Threat Feeds");
         $TopThreatFeeds = QueryCubeJS('{"measures":["PortunusAggSecurity.feednameCount"],"dimensions":["PortunusAggSecurity.feed_name"],"timeDimensions":[{"dimension":"PortunusAggSecurity.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggSecurity.type","operator":"equals","values":["2"]},{"member":"PortunusAggSecurity.severity","operator":"equals","values":["High"]}],"limit":"10","ungrouped":false}');
         if (isset($TopThreatFeeds->result->data)) {
             $TopThreatFeedsSS = IOFactory::load($FilesDir.'/reports/report-'.$UUID.'/ppt/embeddings/Microsoft_Excel_Worksheet.xlsx');
@@ -44,9 +59,9 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
             $TopThreatFeedsW = IOFactory::createWriter($TopThreatFeedsSS, 'Xlsx');
             $TopThreatFeedsW->save($FilesDir.'/reports/report-'.$UUID.'/ppt/embeddings/Microsoft_Excel_Worksheet.xlsx');
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // Top detected properties
+        $Progress = writeProgress($UUID,$Progress,"Getting Threat Properties");
         $TopDetectedProperties = QueryCubeJS('{"measures":["PortunusDnsLogs.tpropertyCount"],"dimensions":["PortunusDnsLogs.tproperty"],"timeDimensions":[{"dimension":"PortunusDnsLogs.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusDnsLogs.type","operator":"equals","values":["2"]},{"member":"PortunusDnsLogs.feed_name","operator":"notEquals","values":["Public_DOH","public-doh","Public_DOH_IP","public-doh-ip"]},{"member":"PortunusDnsLogs.severity","operator":"notEquals","values":["Low","Info"]}],"limit":"10","ungrouped":false}');
         if (isset($TopDetectedProperties->result->data)) {
             $TopDetectedPropertiesSS = IOFactory::load($FilesDir.'/reports/report-'.$UUID.'/ppt/embeddings/Microsoft_Excel_Worksheet1.xlsx');
@@ -60,9 +75,9 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
             $TopDetectedPropertiesW = IOFactory::createWriter($TopDetectedPropertiesSS, 'Xlsx');
             $TopDetectedPropertiesW->save($FilesDir.'/reports/report-'.$UUID.'/ppt/embeddings/Microsoft_Excel_Worksheet1.xlsx');
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // Content filtration
+        $Progress = writeProgress($UUID,$Progress,"Getting Content Filters");
         $ContentFiltration = QueryCubeJS('{"measures":["PortunusAggWebcontent.categoryCount"],"dimensions":["PortunusAggWebcontent.category"],"timeDimensions":[{"dimension":"PortunusAggWebcontent.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[],"limit":"10","ungrouped":false}');
         if (isset($ContentFiltration->result->data)) {
             $ContentFiltrationSS = IOFactory::load($FilesDir.'/reports/report-'.$UUID.'/ppt/embeddings/Microsoft_Excel_Worksheet2.xlsx');
@@ -76,9 +91,9 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
             $ContentFiltrationW = IOFactory::createWriter($ContentFiltrationSS, 'Xlsx');
             $ContentFiltrationW->save($FilesDir.'/reports/report-'.$UUID.'/ppt/embeddings/Microsoft_Excel_Worksheet2.xlsx');
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // Insight Distribution by Threat Type - Sheet 3
+        $Progress = writeProgress($UUID,$Progress,"Getting SOC Insight Threat Types");
         $InsightDistribution = QueryCubeJS('{"measures":["InsightsAggregated.count"],"dimensions":["InsightsAggregated.threatType"],"filters":[{"member":"InsightsAggregated.insightStatus","operator":"equals","values":["Active"]}]}');
         if (isset($InsightDistribution->result->data)) {
             $InsightDistributionSS = IOFactory::load($FilesDir.'/reports/report-'.$UUID.'/ppt/embeddings/Microsoft_Excel_Worksheet3.xlsx');
@@ -92,9 +107,9 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
             $InsightDistributionW = IOFactory::createWriter($InsightDistributionSS, 'Xlsx');
             $InsightDistributionW->save($FilesDir.'/reports/report-'.$UUID.'/ppt/embeddings/Microsoft_Excel_Worksheet3.xlsx');
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // Threat Types (Lookalikes) - Sheet 4
+        $Progress = writeProgress($UUID,$Progress,"Getting Lookalike Threats");
         $LookalikeThreatCountUri = urlencode('/api/atclad/v1/lookalike_threat_counts?_filter=detected_at>="'.$StartDimension.'" and detected_at<="'.$EndDimension.'"');
         $LookalikeThreatCounts = QueryCSP("get",$LookalikeThreatCountUri);
         if (isset($LookalikeThreatCounts->results)) {
@@ -124,18 +139,10 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
             $LookalikeThreatCountsW = IOFactory::createWriter($LookalikeThreatCountsSS, 'Xlsx');
             $LookalikeThreatCountsW->save($FilesDir.'/reports/report-'.$UUID.'/ppt/embeddings/Microsoft_Excel_Worksheet4.xlsx');
         }
-        $Progress = writeProgress($UUID,$Progress);
-
-        // Rebuild Powerpoint Template Zip
-        compressZip($FilesDir.'/reports/report-'.$UUID.'-extracted.pptx',$FilesDir.'/reports/report-'.$UUID);
-        $Progress = writeProgress($UUID,$Progress);
-
-        // Cleanup Extracted Zip
-        rmdirRecursive($FilesDir.'/reports/report-'.$UUID);
-        $Progress = writeProgress($UUID,$Progress);
 
         // ** Reusable Metrics ** //
         // DNS Firewall Activity - Used on Slides 2, 5 & 6
+        $Progress = writeProgress($UUID,$Progress,"Getting DNS Firewall Event Criticality");
         $DNSFirewallActivity = QueryCubeJS('{"measures":["PortunusAggSecurity.severityCount"],"dimensions":["PortunusAggSecurity.severity"],"timeDimensions":[{"dimension":"PortunusAggSecurity.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggSecurity.type","operator":"equals","values":["2","3"]},{"member":"PortunusAggSecurity.severity","operator":"equals","values":["High","Medium","Low"]}],"limit":"3","ungrouped":false}');
         if (isset($DNSFirewallActivity->result)) {
             $HighId = array_search('High', array_column($DNSFirewallActivity->result->data, 'PortunusAggSecurity.severity'));
@@ -159,18 +166,18 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
         $HighPerc = $HighEventsCount * $HMLP;
         $MediumPerc = $MediumEventsCount * $HMLP;
         $LowPerc = $LowEventsCount * $HMLP;
-        $Progress = writeProgress($UUID,$Progress);
 
         // Total DNS Activity - Used on Slides 6 & 9
+        $Progress = writeProgress($UUID,$Progress,"Getting DNS Activity");
         $DNSActivity = QueryCubeJS('{"measures":["PortunusAggInsight.requests"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"equals","values":["1"]}],"limit":"1","ungrouped":false}');
         if (isset($DNSActivity->result->data[0])) {
             $DNSActivityCount = $DNSActivity->result->data[0]->{'PortunusAggInsight.requests'};
         } else {
             $DNSActivityCount = 0;
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // Lookalike Domains - Used on Slides 5, 6 & 24
+        $Progress = writeProgress($UUID,$Progress,"Getting Lookalike Domain Counts");
         $LookalikeDomainCounts = QueryCSP("get","api/atcfw/v1/lookalike_domain_counts");
         if (isset($LookalikeDomainCounts->results->count_total)) { $LookalikeTotalCount = $LookalikeDomainCounts->results->count_total; } else { $LookalikeTotalCount = 0; }
         if (isset($LookalikeDomainCounts->results->percentage_increase_total)) { $LookalikeTotalPercentage = $LookalikeDomainCounts->results->percentage_increase_total; } else { $LookalikeTotalPercentage = 0; }
@@ -178,9 +185,9 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
         if (isset($LookalikeDomainCounts->results->percentage_increase_custom)) { $LookalikeCustomPercentage = $LookalikeDomainCounts->results->percentage_increase_custom; } else { $LookalikeCustomPercentage = 0; }
         if (isset($LookalikeDomainCounts->results->count_threats)) { $LookalikeThreatCount = $LookalikeDomainCounts->results->count_threats; } else { $LookalikeThreatCount = 0; }
         if (isset($LookalikeDomainCounts->results->percentage_increase_threats)) { $LookalikeThreatPercentage = $LookalikeDomainCounts->results->percentage_increase_threats; } else { $LookalikeThreatPercentage = 0; }
-        $Progress = writeProgress($UUID,$Progress);
 
         // SOC Insights - Used on Slides 15 & 28
+        $Progress = writeProgress($UUID,$Progress,"Getting SOC Insight Threat Criticality");
         $SOCInsights = QueryCubeJS('{"measures":["InsightsAggregated.count","InsightsAggregated.mostRecentAt","InsightsAggregated.startedAtMin"],"dimensions":["InsightsAggregated.priorityText"],"filters":[{"member":"InsightsAggregated.insightStatus","operator":"equals","values":["Active"]}],"timezone":"UTC"}');
         if (isset($SOCInsights->result)) {
             $InfoInsightsId = array_search('INFO', array_column($SOCInsights->result->data, 'InsightsAggregated.priorityText'));
@@ -197,45 +204,45 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
         if (isset($MediumInsightsId) AND $MediumInsightsId !== false) {$MediumInsights = $SOCInsights->result->data[$MediumInsightsId]->{'InsightsAggregated.count'};} else {$MediumInsights = 0;}
         if (isset($HighInsightsId) AND $HighInsightsId !== false) {$HighInsights = $SOCInsights->result->data[$HighInsightsId]->{'InsightsAggregated.count'};} else {$HighInsights = 0;}
         if (isset($CriticalInsightsId) AND $CriticalInsightsId !== false) {$CriticalInsights = $SOCInsights->result->data[$CriticalInsightsId]->{'InsightsAggregated.count'};} else {$CriticalInsights = 0;}
-        $Progress = writeProgress($UUID,$Progress);
 
         // Security Activity
+        $Progress = writeProgress($UUID,$Progress,"Getting Security Activity");
         $SecurityEvents = QueryCubeJS('{"measures":["PortunusAggInsight.requests"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"contains","values":["2","3"]}],"limit":"1","ungrouped":false}');
         if (isset($SecurityEvents->result->data[0])) {
             $SecurityEventsCount = $SecurityEvents->result->data[0]->{'PortunusAggInsight.requests'};
         } else {
             $SecurityEventsCount = 0;
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // Data Exfiltration Events
+        $Progress = writeProgress($UUID,$Progress,"Getting Data Exfiltration Events");
         $DataExfilEvents = QueryCubeJS('{"measures":["PortunusAggInsight.requests"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"equals","values":["4"]},{"member":"PortunusAggInsight.tclass","operator":"equals","values":["TI-DNST"]}],"ungrouped":false}');
         if (isset($DataExfilEvents->result->data[0])) {
             $DataExfilEventsCount = $DataExfilEvents->result->data[0]->{'PortunusAggInsight.requests'};
         } else {
             $DataExfilEventsCount = 0;
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // Zero Day DNS Events
+        $Progress = writeProgress($UUID,$Progress,"Getting Zero Day DNS Events");
         $ZeroDayDNSEvents = QueryCubeJS('{"measures":["PortunusAggInsight.requests"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"equals","values":["2","3"]},{"member":"PortunusAggInsight.tclass","operator":"equals","values":["Zero Day DNS"]}],"ungrouped":false}');
         if (isset($ZeroDayDNSEvents->result->data[0])) {
             $ZeroDayDNSEventsCount = $ZeroDayDNSEvents->result->data[0]->{'PortunusAggInsight.requests'};
         } else {
             $ZeroDayDNSEventsCount = 0;
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // Suspicious Domains
+        $Progress = writeProgress($UUID,$Progress,"Getting Suspicious Domain Events");
         $SuspiciousEvents = QueryCubeJS('{"measures":["PortunusAggInsight.requests"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"equals","values":["2"]},{"member":"PortunusAggInsight.tclass","operator":"equals","values":["Suspicious"]}],"ungrouped":false}');
         if (isset($SuspiciousEvents->result->data[0])) {
             $SuspiciousEventsCount = $SuspiciousEvents->result->data[0]->{'PortunusAggInsight.requests'};
         } else {
             $SuspiciousEventsCount = 0;
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // High Risk Websites
+        $Progress = writeProgress($UUID,$Progress,"Getting High Risk Website Events");
         $HighRiskWebsites = QueryCubeJS('{"timeDimensions":[{"dimension":"PortunusAggWebContentDiscovery.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"measures":["PortunusAggWebContentDiscovery.count","PortunusAggWebContentDiscovery.deviceCount"],"dimensions":["PortunusAggWebContentDiscovery.domain_category"],"order":{"PortunusAggWebContentDiscovery.count":"desc"},"filters":[{"member":"PortunusAggWebContentDiscovery.domain_category","operator":"equals","values":["Risky Activity","Suspicious and Malicious Software","Uncategorized","Adult","Abortion","Abortion Pro Choice","Abortion Pro Life","Child Inappropriate","Gambling","Gay","Lingerie","Nudity","Pornography","Profanity","R-Rated","Sex & Erotic","Sex Education","Tobacco","Anonymizer","Criminal Skills","Self Harm","Criminal Activities - Other","Illegal Drugs","Marijuana","Child Abuse Images","Hacking","Hate Speech","Piracy & Copyright Theft","Torrent Repository","Terrorism","Peer-to-Peer","Violence","Weapons","School Cheating","Ad Fraud","Botnet","Command and Control Centers","Compromised & Links To Malware","Malware Call-Home","Malware Distribution Point","Phishing/Fraud","Spam URLs","Spyware & Questionable Software","Cryptocurrency Mining","Sexuality","Parked & For Sale Domains"]}]}');
         if (isset($HighRiskWebsites->result->data)) {
             $HighRiskWebsiteCount = array_sum(array_column($HighRiskWebsites->result->data, 'PortunusAggWebContentDiscovery.count'));
@@ -244,127 +251,249 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
             $HighRiskWebsiteCount = 0;
             $HighRiskWebCategoryCount = 0;
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // DNS over HTTPS
+        $Progress = writeProgress($UUID,$Progress,"Getting DoH Events");
         $DOHEvents = QueryCubeJS('{"measures":["PortunusAggInsight.requests"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"equals","values":["2"]},{"member":"PortunusAggInsight.tproperty","operator":"equals","values":["DoHService"]}],"ungrouped":false}');
         if (isset($DOHEvents->result->data[0])) {
             $DOHEventsCount = $DOHEvents->result->data[0]->{'PortunusAggInsight.requests'};
         } else {
             $DOHEventsCount = 0;
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // Newly Observed Domains
+        $Progress = writeProgress($UUID,$Progress,"Getting Newly Observed Domain Events");
         $NODEvents = QueryCubeJS('{"measures":["PortunusAggInsight.requests"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"equals","values":["2"]},{"member":"PortunusAggInsight.tproperty","operator":"equals","values":["NewlyObservedDomains"]}],"ungrouped":false}');
         if (isset($NODEvents->result->data[0])) {
             $NODEventsCount = $NODEvents->result->data[0]->{'PortunusAggInsight.requests'};
         } else {
             $NODEventsCount = 0;
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // Domain Generation Algorithms
+        $Progress = writeProgress($UUID,$Progress,"Getting DGA Events");
         $DGAEvents = QueryCubeJS('{"measures":["PortunusAggInsight.requests"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"or":[{"member":"PortunusAggInsight.tproperty","operator":"equals","values":["suspicious_rdga","DGA"]},{"member":"PortunusAggInsight.tclass","operator":"equals","values":["DGA","MalwareC2DGA"]}]},{"member":"PortunusAggInsight.type","operator":"equals","values":["2","3"]}],"ungrouped":false}');
         if (isset($DGAEvents->result->data[0])) {
             $DGAEventsCount = $DGAEvents->result->data[0]->{'PortunusAggInsight.requests'};
         } else {
             $DGAEventsCount = 0;
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // Unique Applications
+        $Progress = writeProgress($UUID,$Progress,"Getting list of Unique Applications");
         $UniqueApplications = QueryCubeJS('{"measures":["PortunusAggAppDiscovery.requests"],"dimensions":["PortunusAggAppDiscovery.app_name","PortunusAggAppDiscovery.app_approval"],"timeDimensions":[{"dimension":"PortunusAggAppDiscovery.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggAppDiscovery.app_name","operator":"set"},{"member":"PortunusAggAppDiscovery.app_name","operator":"notEquals","values":[""]}],"order":{}}');
         if (isset($UniqueApplications->result->data)) {
             $UniqueApplicationsCount = count($UniqueApplications->result->data);
         } else {
             $UniqueApplicationsCount = 0;
         }
-        $Progress = writeProgress($UUID,$Progress);
 
-        // Threat Actors
-        $ThreatActors = QueryCubeJS('{"timeDimensions":[{"dateRange":["'.$StartDimension.'","'.$EndDimension.'"],"dimension":"PortunusAggIPSummary.timestamp","granularity":null}],"measures":["PortunusAggIPSummary.count"],"dimensions":["PortunusAggIPSummary.threat_indicator","PortunusAggIPSummary.actor_id"],"filters":[{"and":[{"member":"PortunusAggIPSummary.threat_indicator","operator":"set"},{"member":"PortunusAggIPSummary.actor_id","operator":"set"}]}],"order":{"PortunusAggIPSummary.timestampMax":"desc"}}');
-        if (isset($ThreatActors->result->data)) {
-            $ThreatActorsCount = count(array_unique(array_column($ThreatActors->result->data, 'PortunusAggIPSummary.actor_id')));
+        // Threat Actors Metrics
+        $Progress = writeProgress($UUID,$Progress,"Getting Threat Actor Metrics");
+        $ThreatActors = GetB1ThreatActors($StartDateTime,$EndDateTime);
+        if (isset($ThreatActors)) {
+            $Progress = writeProgress($UUID,$Progress,"Getting Threat Actor Information (This may take a moment)");
+            $ThreatActorsCount = count(array_unique(array_column($ThreatActors, 'PortunusAggIPSummary.actor_id')));
+            $ThreatActorInfo = GetB1ThreatActorsById($ThreatActors);
+            if (isset($ThreatActorInfo->error)) {
+                $ThreatActorInfo = array();
+                $ThreatActorSlideCount = 0;
+            } else {
+                $ThreatActorSlideCount = count($ThreatActorInfo);
+            }
         } else {
             $ThreatActorsCount = 0;
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // Threat Activity
+        $Progress = writeProgress($UUID,$Progress,"Getting Threat Activity");
         $ThreatActivityEvents = QueryCubeJS('{"measures":["PortunusAggInsight.threatCount"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"equals","values":["2"]},{"member":"PortunusAggInsight.severity","operator":"equals","values":["High","Medium","Low"]},{"member":"PortunusAggInsight.threat_indicator","operator":"notEquals","values":[""]}],"limit":"1","ungrouped":false}');
         if (isset($ThreatActivityEvents->result->data[0])) {
             $ThreatActivityEventsCount = $ThreatActivityEvents->result->data[0]->{'PortunusAggInsight.threatCount'};
         } else {
             $ThreatActivityEventsCount = 0;
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // DNS Firewall
+        $Progress = writeProgress($UUID,$Progress,"Getting DNS Firewall Activity");
         $DNSFirewallEvents = QueryCubeJS('{"measures":["PortunusAggInsight.requests"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"and":[{"member":"PortunusAggInsight.type","operator":"equals","values":["2"]},{"or":[{"member":"PortunusAggInsight.severity","operator":"equals","values":["High","Medium","Low"]},{"and":[{"member":"PortunusAggInsight.severity","operator":"equals","values":["Info"]},{"member":"PortunusAggInsight.policy_action","operator":"equals","values":["Block","Log"]}]}]},{"member":"PortunusAggInsight.confidence","operator":"equals","values":["High","Medium","Low"]}]}],"limit":"1","ungrouped":false}');
         if (isset($DNSFirewallEvents->result->data[0])) {
             $DNSFirewallEventsCount = $DNSFirewallEvents->result->data[0]->{'PortunusAggInsight.requests'};
         } else {
             $DNSFirewallEventsCount = 0;
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // Web Content
+        $Progress = writeProgress($UUID,$Progress,"Getting Web Content Events");
         $WebContentEvents = QueryCubeJS('{"measures":["PortunusAggWebcontent.requests"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggWebcontent.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggWebcontent.type","operator":"equals","values":["3"]},{"member":"PortunusAggWebcontent.category","operator":"notEquals","values":[null]}],"limit":"1","ungrouped":false}');
         if (isset($WebContentEvents->result->data[0])) {
             $WebContentEventsCount = $WebContentEvents->result->data[0]->{'PortunusAggWebcontent.requests'};
         } else {
             $WebContentEventsCount = 0;
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // Device Count
+        $Progress = writeProgress($UUID,$Progress,"Getting Device Count");
         $Devices = QueryCubeJS('{"measures":["PortunusAggInsight.deviceCount"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"contains","values":["2","3"]},{"member":"PortunusAggInsight.severity","operator":"contains","values":["High","Medium","Low"]}],"limit":"1","ungrouped":false}');
         if (isset($Devices->result->data[0])) {
             $DeviceCount = $Devices->result->data[0]->{'PortunusAggInsight.deviceCount'};
         } else {
             $DeviceCount = 0;
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // User Count
+        $Progress = writeProgress($UUID,$Progress,"Getting User Count");
         $Users = QueryCubeJS('{"measures":["PortunusAggInsight.userCount"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"contains","values":["2","3"]}],"limit":"1","ungrouped":false}');
         if (isset($Users->result->data[0])) {
             $UserCount = $Users->result->data[0]->{'PortunusAggInsight.userCount'};
         } else {
             $UserCount = 0;
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // Threat Insight Count
+        $Progress = writeProgress($UUID,$Progress,"Getting Threat Insight Count");
         $ThreatInsight = QueryCubeJS('{"measures":[],"dimensions":["PortunusDnsLogs.tproperty"],"timeDimensions":[{"dimension":"PortunusDnsLogs.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusDnsLogs.type","operator":"equals","values":["4"]}],"limit":"10000","ungrouped":false}'); // Threat Insight
         if (isset($ThreatInsight->result->data)) {
             $ThreatInsightCount = count($ThreatInsight->result->data);
         } else {
             $ThreatInsightCount = 0;
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // Threat View Count
+        $Progress = writeProgress($UUID,$Progress,"Getting Threat View Count");
         $ThreatView = QueryCubeJS('{"measures":["PortunusAggInsight.tpropertyCount"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggInsight.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggInsight.type","operator":"equals","values":["2"]}],"limit":"1","ungrouped":false}');
         if (isset($ThreatView->result->data[0])) {
             $ThreatViewCount = $ThreatView->result->data[0]->{'PortunusAggInsight.tpropertyCount'};
         } else {
             $ThreatViewCount = 0;
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // Source Count
+        $Progress = writeProgress($UUID,$Progress,"Getting Sources Count");
         $Sources = QueryCubeJS('{"measures":["PortunusAggSecurity.networkCount"],"dimensions":[],"timeDimensions":[{"dimension":"PortunusAggSecurity.timestamp","dateRange":["'.$StartDimension.'","'.$EndDimension.'"]}],"filters":[{"member":"PortunusAggSecurity.type","operator":"contains","values":["2","3"]}],"limit":"1","ungrouped":false}');
         if (isset($Sources->result->data[0])) {
             $SourcesCount = $Sources->result->data[0]->{'PortunusAggSecurity.networkCount'};
         } else {
             $SourcesCount = 0;
         }
-        $Progress = writeProgress($UUID,$Progress);
 
         // ** ** //
 
+
+        //
+        // Do Threat Actor Stuff Here ....
+        //
+        $Progress = writeProgress($UUID,$Progress,"Generating Threat Actor Slides");
+        // Calculate the slide position based on above value
+        $ThreatActorSlidePosition = $ThreatActorSlideStart-2;
+
+        // Tag Numbers Start
+        $TagStart = 100;
+
+        // Open PPTX Presentation _rels XML
+        $xml_rels = new DOMDocument('1.0', 'utf-8');
+        $xml_rels->formatOutput = true; 
+        $xml_rels->preserveWhiteSpace = false;
+        $xml_rels->load($FilesDir.'/reports/report-'.$UUID.'/ppt/_rels/presentation.xml.rels');
+        $xml_rels_f = $xml_rels->createDocumentFragment();
+        $xml_rels_fstart = ($xml_rels->getElementsByTagName('Relationship')->length)+50;
+
+        // Open PPTX Presentation XML
+        $xml_pres = new DOMDocument('1.0', 'utf-8');
+        $xml_pres->formatOutput = true;
+        $xml_pres->preserveWhiteSpace = false;
+        $xml_pres->load($FilesDir.'/reports/report-'.$UUID.'/ppt/presentation.xml');
+        $xml_pres_f = $xml_pres->createDocumentFragment();
+        $xml_pres_fstart = 14700;
+
+        // Get Slide Count
+        $SlidesCount = iterator_count(new FilesystemIterator($FilesDir.'/reports/report-'.$UUID.'/ppt/slides'));
+        // Set first slide number
+        $SlideNumber = $SlidesCount++;
+
+        // Build new Threat Actor Slides & Update PPTX Resources
+        foreach  ($ThreatActorInfo as $TAI) {
+            if (($ThreatActorSlideCount - 1) > 0) {
+                $xml_rels_f->appendXML('<Relationship Id="rId'.$xml_rels_fstart.'" Type="http://schemas.openxmlformats.org/officeDocument/2006/relationships/slide" Target="slides/slide'.$SlideNumber.'.xml"/>');
+                $xml_pres_f->appendXML('<p:sldId id="'.$xml_pres_fstart.'" r:id="rId'.$xml_rels_fstart.'"/>');
+    
+                $xml_rels_fstart++;
+                $xml_pres_fstart++;
+    
+                copy($FilesDir.'/reports/report-'.$UUID.'/ppt/slides/slide'.$ThreatActorSlideStart.'.xml',$FilesDir.'/reports/report-'.$UUID.'/ppt/slides/slide'.$SlideNumber.'.xml');
+                copy($FilesDir.'/reports/report-'.$UUID.'/ppt/slides/_rels/slide'.$ThreatActorSlideStart.'.xml.rels',$FilesDir.'/reports/report-'.$UUID.'/ppt/slides/_rels/slide'.$SlideNumber.'.xml.rels');
+            } else {
+                $SlideNumber = $ThreatActorSlideStart;
+            }
+
+            // Update Tag Numbers
+            $TASFile = file_get_contents($FilesDir.'/reports/report-'.$UUID.'/ppt/slides/slide'.$SlideNumber.'.xml');
+            $TASFileString = str_replace('#TATAG00', '#TATAG'.$TagStart, $TASFile);
+            file_put_contents($FilesDir.'/reports/report-'.$UUID.'/ppt/slides/slide'.$SlideNumber.'.xml', $TASFileString);
+
+            $xml_tas = new DOMDocument('1.0', 'utf-8');
+            $xml_tas->formatOutput = true; 
+            $xml_tas->preserveWhiteSpace = false;
+            $xml_tas->load($FilesDir.'/reports/report-'.$UUID.'/ppt/slides/_rels/slide'.$SlideNumber.'.xml.rels');
+
+            foreach ($xml_tas->getElementsByTagName('Relationship') as $element) {
+                // Remove notes references to avoid having to create unneccessary notes resources
+                if ($element->getAttribute('Type') == "http://schemas.openxmlformats.org/officeDocument/2006/relationships/notesSlide") {
+                    $element->remove();
+                }
+                if (array_key_exists($TAI['actor_name'],$KnownActors)) {
+                    $UniqueActor = $KnownActors[$TAI['actor_name']];
+                    // Threat Actor PNG
+                    if ($element->getAttribute('Id') == $ThreatActorPNGID) {
+                        copy($AssetsDir.'/images/Threat Actors/Glow/'.$UniqueActor.'.png',$FilesDir.'/reports/report-'.$UUID.'/ppt/media/'.$UniqueActor.'.png');
+                        $element->setAttribute('Target','../media/'.$UniqueActor.'.png');
+                    }
+                    // Threat Actor SVG
+                    if ($element->getAttribute('Id') == $ThreatActorSVGID) {
+                        copy($AssetsDir.'/images/Threat Actors/Glow/'.$UniqueActor.'.svg',$FilesDir.'/reports/report-'.$UUID.'/ppt/media/'.$UniqueActor.'.svg');
+                        $element->setAttribute('Target','../media/'.$UniqueActor.'.svg');
+                    }
+                }
+            }
+            $xml_tas->save($FilesDir.'/reports/report-'.$UUID.'/ppt/slides/_rels/slide'.$SlideNumber.'.xml.rels');
+            $TagStart += 10;
+            // Iterate slide number
+            $SlideNumber++;
+            $ThreatActorSlideCount--;
+        }
+
+        // Append Elements to Core XML Files
+        $xml_rels->getElementsByTagName('Relationships')->item(0)->appendChild($xml_rels_f);
+        // Append new slides to specific position
+        $xml_pres->getElementsByTagName('sldId')->item($ThreatActorSlidePosition)->after($xml_pres_f);
+        
+        // Save Core XML Files
+        $xml_rels->save($FilesDir.'/reports/report-'.$UUID.'/ppt/_rels/presentation.xml.rels');
+        $xml_pres->save($FilesDir.'/reports/report-'.$UUID.'/ppt/presentation.xml');
+        
+        //
+        // End of Threat Actors
+        //
+       
+
+        // Rebuild Powerpoint Template Zip
+        $Progress = writeProgress($UUID,$Progress,"Stitching Powerpoint Template");
+        compressZip($FilesDir.'/reports/report-'.$UUID.'.pptx',$FilesDir.'/reports/report-'.$UUID);
+
+        // Cleanup Extracted Zip
+        $Progress = writeProgress($UUID,$Progress,"Cleaning up");
+        rmdirRecursive($FilesDir.'/reports/report-'.$UUID);
+
+        // Extract Powerpoint Template Strings
+        // ** Using external library to save re-writing the string replacement functions manually. Will probably pull this in as native code at some point.
+        $Progress = writeProgress($UUID,$Progress,"Extract Powerpoint Strings");
+        $extractor = new BasicExtractor();
+        $mapping = $extractor->extractStringsAndCreateMappingFile(
+            $FilesDir.'/reports/report-'.$UUID.'.pptx',
+            $FilesDir.'/reports/report-'.$UUID.'-extracted.pptx'
+        );
+
+        $Progress = writeProgress($UUID,$Progress,"Injecting Powerpoint Strings");
         ##// Slide 2 / 45 - Title Page & Contact Page
         // Get & Inject Customer Name, Contact Name & Email
         $AccountInfo = QueryCSP("get","v2/current_user/accounts");
@@ -374,7 +503,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
         $mapping = replaceTag($mapping,'#DATE',date("dS F Y"));
         $mapping = replaceTag($mapping,'#NAME',$UserInfo->result->name);
         $mapping = replaceTag($mapping,'#EMAIL',$UserInfo->result->email);
-        $Progress = writeProgress($UUID,$Progress);
 
         ##// Slide 5 - Executive Summary
         $mapping = replaceTag($mapping,'#TAG02',number_abbr($HighEventsCount)); // High-Risk Events
@@ -383,7 +511,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
         $mapping = replaceTag($mapping,'#TAG05',number_abbr($LookalikeThreatCount)); // Lookalike Domains
         $mapping = replaceTag($mapping,'#TAG06',number_abbr($ZeroDayDNSEventsCount)); // Zero Day DNS
         $mapping = replaceTag($mapping,'#TAG07',number_abbr($SuspiciousEventsCount)); // Suspicious Domains
-        $Progress = writeProgress($UUID,$Progress);
 
         ##// Slide 6 - Security Indicator Summary
         $mapping = replaceTag($mapping,'#TAG08',number_abbr($DNSActivityCount)); // DNS Requests
@@ -401,7 +528,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
         $mapping = replaceTag($mapping,'#TAG19',number_abbr($UniqueApplicationsCount)); // Unique Applications
         $mapping = replaceTag($mapping,'#TAG20',number_abbr($HighRiskWebCategoryCount)); // High-Risk Web Categories
         $mapping = replaceTag($mapping,'#TAG21',number_abbr($ThreatActorsCount)); // Threat Actors
-        $Progress = writeProgress($UUID,$Progress);
 
         ##// Slide 9 - Traffic Usage Analysis
         // Total DNS Activity
@@ -418,7 +544,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
         $mapping = replaceTag($mapping,'#TAG30',number_abbr($ThreatActivityEventsCount));
         // Data Exfiltration Incidents
         $mapping = replaceTag($mapping,'#TAG31',number_abbr($DataExfilEventsCount));
-        $Progress = writeProgress($UUID,$Progress);
 
         ##// Slide 15 - Key Insights
         // Insight Severity
@@ -429,7 +554,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
         // Event To Insight Aggregation
         $mapping = replaceTag($mapping,'#TAG36',number_abbr($SecurityEventsCount)); // Events
         $mapping = replaceTag($mapping,'#TAG37',number_abbr($TotalInsights)); // Key Insights
-        $Progress = writeProgress($UUID,$Progress);
 
         ##// Slide 24 - Lookalike Domains
         $mapping = replaceTag($mapping,'#TAG38',number_abbr($LookalikeTotalCount)); // Total Lookalikes
@@ -444,7 +568,6 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
         if ($LookalikeThreatPercentage >= 0){$arrow='↑';} else {$arrow='↓';}
         $mapping = replaceTag($mapping,'#TAG45',$arrow); // Arrow Up/Down
         $mapping = replaceTag($mapping,'#TAG46',number_abbr($LookalikeThreatPercentage)); // Threats Percentage Increase
-        $Progress = writeProgress($UUID,$Progress);
 
         ##// Slide 28 - Security Activities
         $mapping = replaceTag($mapping,'#TAG47',number_abbr($SecurityEventsCount)); // Security Events
@@ -456,21 +579,80 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
         $mapping = replaceTag($mapping,'#TAG53',number_abbr($ThreatInsightCount)); // Threat Insight
         $mapping = replaceTag($mapping,'#TAG54',number_abbr($ThreatViewCount)); // Threat View
         $mapping = replaceTag($mapping,'#TAG55',number_abbr($SourcesCount)); // Sources
-        $Progress = writeProgress($UUID,$Progress);
+
+        ##// Slide 32 -> Onwards - Threat Actors
+        // This is where the Threat Actor Tag replacement occurs
+        // Set Tag Start Number
+        $TagStart = 100;
+        foreach ($ThreatActorInfo as $TAI) {
+            // Get sorted list of observed IOCs not found in Virus Total
+            if (isset($TAI['related_indicators_with_dates'])) {
+                $ObservedIndicators = $TAI['related_indicators_with_dates'];
+                $IndicatorsInVT = [];
+                $IndicatorCount = count($TAI['related_indicators_with_dates']);
+                if ($IndicatorCount > 0) {
+                    foreach ($ObservedIndicators as $OI) {
+                        if (array_key_exists('vt_first_submission_date', json_decode(json_encode($OI), true))) {
+                            $IndicatorsInVT[] = $OI;
+                        }
+                    }
+                }
+                if (count($IndicatorsInVT) > 0) {
+                    // Sort the array based on the time difference
+                    usort($IndicatorsInVT, function($a, $b) {
+                        return calculateVirusTotalDifference($b) <=> calculateVirusTotalDifference($a);
+                    });
+                    $IndicatorsNotInVT = count($ObservedIndicators) - count($IndicatorsInVT);
+                    $ExampleDomain = $IndicatorsInVT[0]->indicator;
+                    $FirstSeen = new DateTime($IndicatorsInVT[0]->te_ik_submitted);
+                    $LastSeen = new DateTime($IndicatorsInVT[0]->te_customer_last_dns_query);
+                    $VTDate = new DateTime($IndicatorsInVT[0]->vt_first_submission_date);
+                    $ProtectedFor = $FirstSeen->diff($LastSeen)->days;
+                    $DaysAhead = 'Discovered '.($ProtectedFor - $LastSeen->diff($VTDate)->days).' days ahead';
+                } else {
+                    $IndicatorsNotInVT = count($ObservedIndicators);
+                    $ExampleDomain = $ObservedIndicators[0]->indicator;
+                    $FirstSeen = new DateTime($ObservedIndicators[0]->te_ik_submitted);
+                    $LastSeen = new DateTime($ObservedIndicators[0]->te_customer_last_dns_query);
+                    $DaysAhead = 'Discovered';
+                    $ProtectedFor = $FirstSeen->diff($LastSeen)->days;
+                }
+            } else {
+                $IndicatorsNotInVT = 'N/A';
+                $ExampleDomain = 'N/A';
+                $FirstSeen = new DateTime('1901-01-01 00:00');
+                $LastSeen = new DateTime('1901-01-01 00:00');
+                $DaysAhead = 'Discovered';
+                $ProtectedFor = 'N/A';
+                $IndicatorCount = 'N/A';
+            }
+
+            $mapping = replaceTag($mapping,'#TATAG'.$TagStart.'01',ucwords($TAI['actor_name'])); // Threat Actor Name
+            $mapping = replaceTag($mapping,'#TATAG'.$TagStart.'02',$TAI['actor_description']); // Threat Actor Description
+            $mapping = replaceTag($mapping,'#TATAG'.$TagStart.'03',$IndicatorCount); // Number of Observed IOCs
+            $mapping = replaceTag($mapping,'#TATAG'.$TagStart.'04',$IndicatorsNotInVT); // Number of Observed IOCs not found in Virus Total
+            $mapping = replaceTag($mapping,'#TATAG'.$TagStart.'05',$TAI['related_count']); // Number of Related IOCs
+            $mapping = replaceTag($mapping,'#TATAG'.$TagStart.'06',$DaysAhead); // Discovered X Days Ahead
+            $mapping = replaceTag($mapping,'#TATAG'.$TagStart.'07',$FirstSeen->format('d/m/y')); // First Detection Date
+            $mapping = replaceTag($mapping,'#TATAG'.$TagStart.'08',$LastSeen->format('d/m/y')); // Last Detection Date
+            $mapping = replaceTag($mapping,'#TATAG'.$TagStart.'09',$ProtectedFor); // Protected X Days
+            $mapping = replaceTag($mapping,'#TATAG'.$TagStart.'10',$ExampleDomain); // Example Domain
+            $TagStart += 10;
+        }
 
         // Rebuild Powerpoint
         // ** Using external library to save re-writing the string replacement functions manually. Will probably pull this in as native code at some point.
+        $Progress = writeProgress($UUID,$Progress,"Rebuilding Powerpoint Template");
         $injector = new BasicInjector();
         $injector->injectMappingAndCreateNewFile(
             $mapping,
             $FilesDir.'/reports/report-'.$UUID.'-extracted.pptx',
             $FilesDir.'/reports/report-'.$UUID.'.pptx'
         );
-        $Progress = writeProgress($UUID,$Progress);
 
         // Cleanup
+        $Progress = writeProgress($UUID,$Progress,"Final Cleanup");
         unlink($FilesDir.'/reports/report-'.$UUID.'-extracted.pptx');
-        $Progress = writeProgress($UUID,$Progress);
 
         $Status = 'Success';
     } else {
@@ -490,10 +672,14 @@ function generateSecurityReport($StartDateTime,$EndDateTime,$Realm,$UUID) {
     return $response;
 }
 
-function writeProgress($id,$Count) {
+function writeProgress($id,$Count,$Action = "") {
     $Count++;
     $myfile = fopen(__DIR__.'/../../files/reports/report-'.$id.'.progress', "w") or die("Unable to save progress file");
-    fwrite($myfile, $Count);
+    $Progress = json_encode(array(
+        'Count' => $Count,
+        'Action' => $Action
+    ));
+    fwrite($myfile, $Progress);
     fclose($myfile);
     return $Count;
 }
@@ -502,14 +688,32 @@ function getProgress($id,$Total) {
     $ProgressFile = __DIR__.'/../../files/reports/report-'.$id.'.progress';
     if (file_exists($ProgressFile)) {
         $myfile = fopen($ProgressFile, "r") or die("0");
-        $Current = fread($myfile,filesize($ProgressFile));
-        return (100 / $Total) * $Current;
+        $Current = json_decode(fread($myfile,filesize($ProgressFile)));
+        return array(
+            'Progress' => (100 / $Total) * $Current->Count,
+            'Action' => $Current->Action.'..'
+        );
     } else {
-        die("0");
+        return array(
+            'Progress' => 0,
+            'Action' => 'Starting..'
+        );
     }
 }
 
 function getReportFiles() {
     $files = array_diff(scandir(__DIR__.'/../../files/reports/'),array('.', '..','placeholder.txt'));
     return $files;
-  }
+}
+
+function calculateProtectedDifference($te_ik_submitted,$te_customer_last_dns_query) {
+    $submitted = new DateTime($te_ik_submitted);
+    $lastQuery = new DateTime($te_customer_last_dns_query);
+    return $lastQuery->getTimestamp() - $submitted->getTimestamp();
+}
+
+function calculateVirusTotalDifference($obj) {
+    $submitted = new DateTime($obj->te_ik_submitted);
+    $vtsubmitted = new DateTime($obj->vt_first_submission_date);
+    return $vtsubmitted->getTimestamp() - $submitted->getTimestamp();
+}